requirements.txt
# Core Dependencies
google-generativeai==0.8.3
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0

# Data Collection
arxiv==2.1.0
feedparser==6.0.10
duckduckgo-search==4.1.1
huggingface-hub==0.20.2
kaggle==1.5.16

# Graph Database (Production)
neo4j==5.16.0

# Vector Database (Production)
qdrant-client==1.7.3

# ML & Embeddings
sentence-transformers==3.3.1
torch==2.1.1

# Development Fallbacks
networkx==3.2.1
faiss-cpu==1.7.4

# Visualization
pyvis==0.3.2

# Data Processing
PyPDF2==3.0.1
pandas==2.1.3
numpy==1.26.2

# Scheduling
schedule==1.2.0
apscheduler==3.10.4

# Logging & Monitoring
structlog==23.2.0

# Web Framework (for future UI)
fastapi==0.109.0
uvicorn==0.27.0

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.3
